add support for fullwidth characters
