+++ index.js
@@ -1,3 +1,4 @@
+var util = require("./util")
 module.exports = blit
 
 // Applies 'lines' to 'screen' at coordinates x/y
@@ -23,46 +24,11 @@ function blit (screen, lines, x, y) {
 // String, String -> String
 function mergeString (src, string, x) {
   var res = src
-  var extraCharsNeeded = (x + strlenAnsi(string)) - strlenAnsi(src)
+  var extraCharsNeeded = (x + util.strlenAnsi(string)) - util.strlenAnsi(src)
   if (extraCharsNeeded > 0) {
     res += (new Array(extraCharsNeeded).fill(' ')).join('')
   }
 
-  return sliceAnsi(res, 0, x) + string + sliceAnsi(res, x + strlenAnsi(string))
+  return util.sliceAnsi(res, 0, x) + string + util.sliceAnsi(res, x + util.strlenAnsi(string))
 }
 
-// Like String#slice, but taking ANSI codes into account
-function sliceAnsi (str, from, to) {
-  var len = 0
-  var insideCode = false
-  var res = ''
-  to = (to === undefined) ? str.length : to
-
-  for (var i=0; i < str.length; i++) {
-    var chr = str.charAt(i)
-    if (chr === '\033') insideCode = true
-    if (!insideCode) len++
-    if (chr === 'm' && insideCode) insideCode = false
-
-    if (len > from && len <= to) {
-      res += chr
-    }
-  }
-
-  return res
-}
-
-// Length of 'str' sans ANSI codes
-function strlenAnsi (str) {
-  var len = 0
-  var insideCode = false
-
-  for (var i=0; i < str.length; i++) {
-    var chr = str.charAt(i)
-    if (chr === '\033') insideCode = true
-    if (!insideCode) len++
-    if (chr === 'm' && insideCode) insideCode = false
-  }
-
-  return len
-}
