+++ package.json
@@ -13,7 +13,9 @@
     "lint": "standard"
   },
   "keywords": [],
-  "dependencies": {},
+  "dependencies": {
+    "is-fullwidth-code-point": "^3.0.0"
+  },
   "devDependencies": {
     "standard": "~10.0.0"
   },
